+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 * Fix Strings.ChrW conversion [#924](https://github.com/icsharpcode/CodeConverter/issues/924)
 * Fix mapping for error trivia added by converter. [#917](https://github.com/icsharpcode/CodeConverter/issues/917)
 * Fix CType expressions for conversions from String to Enums [#918](https://github.com/icsharpcode/CodeConverter/issues/918)
+* Only pull loop variables out when it's needed [#915](https://github.com/icsharpcode/CodeConverter/issues/915)
 
 ### C# -> VB